+++ src/main/java/org/terracotta/ehcache/testing/objectgenerator/ObjectGenerator.java
@@ -2,6 +2,6 @@ package org.terracotta.ehcache.testing.objectgenerator;
 
 public interface ObjectGenerator {
 
-  Object generate(int seed);
+  Object generate(long seed);
 
 }
